requirements.txt
aiogram>=3.0
asyncpg==0.29.0  # Асинхронный PostgreSQL-клиент
python-dotenv==1.0.0  # Для загрузки переменных окружения
sqlalchemy
phonenumbers